lwnbd: move IOP specific code in platform-ps2.h
