+++ modules/network/lwnbdsvr/lwnbd.h
@@ -41,11 +41,9 @@
 #ifndef LWIP_HDR_APPS_NBD_SERVER_H
 #define LWIP_HDR_APPS_NBD_SERVER_H
 
+#include <stdio.h>
 #include "nbd-protocol.h"
 #include "nbd_opts.h"
-
-#include <stdint.h>
-#include <stdio.h>
 
 #define LOG(format, args...) \
     printf(APP_NAME ": " format, ##args)
@@ -57,68 +55,6 @@
     do {                  \
     } while (0)
 #endif
-
-//#include "lwip/apps/nbd_opts.h"
-//#include "lwip/err.h"
-//#include "lwip/pbuf.h"
-//#include "lwip/mem.h"
-
-#ifdef __linux__
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <endian.h>
-#include <unistd.h>
-
-#include <assert.h> //todo: move in .c
-
-typedef signed char err_t;
-//TODO : manage endianess
-#define htonll(x) htobe64(x)
-#define ntohll(x) be64toh(x)
-#endif
-
-#ifdef _IOP
-#include <ps2ip.h>
-#include <sysclib.h>
-
-#define assert(expr) \
-    ((expr) ||       \
-     dbgprintf(F_NUM, __LINE__))
-
-//#include <errno.h>
-//#include <malloc.h>
-
-//why not provide lwip/sockets.h ?
-// #define send(a, b, c, d) lwip_send(a, b, c, d)
-#define close(x) lwip_close(x)
-
-//TODO: Missing <byteswap.h> in PS2SDK
-// pickup from https://gist.github.com/jtbr/7a43e6281e6cca353b33ee501421860c
-static inline uint64_t bswap64(uint64_t x)
-{
-    return (((x & 0xff00000000000000ull) >> 56) | ((x & 0x00ff000000000000ull) >> 40) | ((x & 0x0000ff0000000000ull) >> 24) | ((x & 0x000000ff00000000ull) >> 8) | ((x & 0x00000000ff000000ull) << 8) | ((x & 0x0000000000ff0000ull) << 24) | ((x & 0x000000000000ff00ull) << 40) | ((x & 0x00000000000000ffull) << 56));
-}
-
-//TODO: Missing in PS2SK's "common/include/tcpip.h"
-#if __BIG_ENDIAN__
-#define htonll(x) (x)
-#define ntohll(x) (x)
-#else
-#define htonll(x) bswap64(x)
-#define ntohll(x) bswap64(x)
-#endif
-
-// Missing in stddef.h
-#ifndef offsetof
-#define offsetof(st, m) \
-    ((size_t)((char *)&((st *)0)->m - (char *)0))
-#endif
-
-//TODO: Missing in PS2SK's <stdint.h> , needed for "nbd-protocol.h"
-// https://en.cppreference.com/w/c/types/integer
-#define UINT64_MAX  0xffffffffffffffff
-#define UINT64_C(x) ((x) + (UINT64_MAX - UINT64_MAX))
-#endif /* _IOP */
 
 #ifdef __cplusplus
 extern "C" {
@@ -134,55 +70,51 @@ struct lwnbd_operations;
 typedef struct nbd_context
 {
     struct lwnbd_operations const *vptr;
-
     char export_desc[64];
     char export_name[32];
     uint64_t export_size; /* size of export in byte */
     uint16_t eflags;      /* per-export flags */
-    uint16_t blocksize;   /* in power of 2 for bit shifting - log2(blocksize) */
+    uint16_t blocksize;
     uint8_t *buffer;
-
 } nbd_context;
 
 struct lwnbd_operations
 {
-
     /**
-   * Close block device handle
-   * @param handle File handle returned by open()
-   */
+    * Close block device handle
+    * @param handle File handle returned by open()
+    */
     //  void (*close)(nbd_context *me);
     /**
-   * Read from block device
-   * @param
-   * @param buffer Target buffer to copy read data to
-   * @param offset Offset in block to copy read data to
-   * @param length Number of blocks to copy to buffer
-   * @returns &gt;= 0: Success; &lt; 0: Error
-   */
+    * Read from block device
+    * @param
+    * @param buffer Target buffer to copy read data to
+    * @param offset Offset in block to copy read data to
+    * @param length Number of blocks to copy to buffer
+    * @returns &gt;= 0: Success; &lt; 0: Error
+    */
     int (*read)(nbd_context const *const me, void *buffer, uint64_t offset, uint32_t length);
     /**
-   * Write to block device
-   * @param me ()
-   * @param buffer Target buffer to copy write data to
-   * @param offset Offset in block to copy write data to
-   * @param length Number of blocks to copy to buffer
-   * @returns &gt;= 0: Success; &lt; 0: Error
-   */
+    * Write to block device
+    * @param me ()
+    * @param buffer Target buffer to copy write data to
+    * @param offset Offset in block to copy write data to
+    * @param length Number of blocks to copy to buffer
+    * @returns &gt;= 0: Success; &lt; 0: Error
+    */
     int (*write)(nbd_context const *const me, void *buffer, uint64_t offset, uint32_t length);
     /**
-   * Flush to block device
-   * @param me ()
-   * @returns &gt;= 0: Success; &lt; 0: Error
-   */
+    * Flush to block device
+    * @param me ()
+    * @returns &gt;= 0: Success; &lt; 0: Error
+    */
     int (*flush)(nbd_context const *const me);
 };
 
 uint32_t nbd_recv(int s, void *mem, size_t len, int flags);
 int nbd_init(nbd_context **ctx);
 
-
-// in nbd_protocol.c
+//********************* nbd_protocol.c *********************
 //todo: const ctxs
 err_t negotiation_phase(const int client_socket, nbd_context **ctxs, nbd_context **ctx);
 err_t transmission_phase(const int client_socket, nbd_context *ctx);
